Added cmake, Licences and libs.
All the libraries were added and every folder was reorganized in order
to be independent from Arduino IDE and be able to compile everything
through a cmake toolchain, which produces directly the .hex file. The
bootloader is added and compiled as compliance to the GPLv2 licence. The
bootloader chosen is the STK500V2 bootloader which is added to the .hex
file automatically. NIM-Counter code was updated.